+++ R/gg.R
@@ -7,18 +7,43 @@ utils::globalVariables(c('time', 'strata', 'n.risk'))
 #' @param fit Required. survival::survfit() object.
 #' @param times Required. Numeric. One or more time points to calculate
 #' the number at risk for.
+#' @param strata.order Optional. Character. Ordered names of strata factor
+#' levels.
 #' @return An unformatted ggplot2 table showing the number at risk.
 #' @examples
 #' library(survival)
 #'
 #' fit <- survfit(Surv(time, status) ~ trt, data = diabetic)
 #'
-#' ggrisktable(fit, c(0, 10, 20, 30, 40, 50)) +
-#'   theme_risk_black()
+#' ggrisktable(
+#'    fit = fit,
+#'    times = c(0, 10, 20, 30, 40, 50),
+#'    strata.order = c('0', '1')
+#' ) + theme_risk_black()
 #' @export
-ggrisktable <- function (fit, times) {
+ggrisktable <- function (fit = NULL, times = NULL, strata.order = NULL) {
+
+  # Hard stops
+  if (is.null(fit) | class(fit) != 'survfit') stop('No valid fit object provided. [Check: \'fit\']')
+  if (is.null(times) | !is.numeric(times)) stop('No valid time points provided. [Check: \'times\']')
+  if (!is.null(strata.order) & !is.character(strata.order)) stop('Invalid strata order data provided. [Check: \'strata.order\']')
+
+  # Generate risk table and order
+  risk_table <- utile.tools::tabulate_at_risk(fit, times)
+
+  # Reorder strata
+  if (is.character(strata.order))
+    risk_table$strata <- factor(
+      risk_table$strata,
+      levels = unique(c(
+        rev(strata.order[strata.order %in% levels(risk_table$strata)]),
+        levels(risk_table$strata)
+      ))
+    )
+
+  # Return plotted table
   ggplot2::ggplot(
-    utile.tools::tabulate_at_risk(fit, times),
+    risk_table,
     ggplot2::aes(x = time, y = strata, label = n.risk)
   ) + ggplot2::geom_text()
 }