+++ NEWS.md
@@ -1,3 +1,10 @@
+# utile.visuals 0.2.1
+* `append_table()`: Legend extraction can now be toggled with the new `extract.legend` option. Extraction may not be a desired behavior if the legend is already embedded within plot area.
+* `ggrisktable()`:
+  - New option `strata.order` added to allow reordering of strata in final table.
+  - Now includes hard stops for invalid option data.
+* Added `connect_origin` for connecting KM curves to a plot origin.
+
 # utile.visuals 0.2.0
 * Added `ggrisktable()` for creating a ggplot2 risk table from a survival::survfit() model.
 * Added `theme_risk_black()` and `theme_risk_white()` for risk table formatting.